# wsgi-scripts/swagger_server/models/training_info.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.body2 import Body2  # noqa: F401,E501
from swagger_server.models.training_info_training_info import TrainingInfoTrainingInfo  # noqa: F401,E501
from swagger_server import util


class TrainingInfo(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, trainings: Body2=None, training_info: List[TrainingInfoTrainingInfo]=None):  # noqa: E501
        """TrainingInfo - a model defined in Swagger

        :param trainings: The trainings of this TrainingInfo.  # noqa: E501
        :type trainings: Body2
        :param training_info: The training_info of this TrainingInfo.  # noqa: E501
        :type training_info: List[TrainingInfoTrainingInfo]
        """
        self.swagger_types = {
            'trainings': Body2,
            'training_info': List[TrainingInfoTrainingInfo]
        }

        self.attribute_map = {
            'trainings': 'Trainings',
            'training_info': 'TrainingInfo'
        }
        self._trainings = trainings
        self._training_info = training_info

    @classmethod
    def from_dict(cls, dikt) -> 'TrainingInfo':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The TrainingInfo of this TrainingInfo.  # noqa: E501
        :rtype: TrainingInfo
        """
        return util.deserialize_model(dikt, cls)

    @property
    def trainings(self) -> Body2:
        """Gets the trainings of this TrainingInfo.


        :return: The trainings of this TrainingInfo.
        :rtype: Body2
        """
        return self._trainings

    @trainings.setter
    def trainings(self, trainings: Body2):
        """Sets the trainings of this TrainingInfo.


        :param trainings: The trainings of this TrainingInfo.
        :type trainings: Body2
        """

        self._trainings = trainings

    @property
    def training_info(self) -> List[TrainingInfoTrainingInfo]:
        """Gets the training_info of this TrainingInfo.


        :return: The training_info of this TrainingInfo.
        :rtype: List[TrainingInfoTrainingInfo]
        """
        return self._training_info

    @training_info.setter
    def training_info(self, training_info: List[TrainingInfoTrainingInfo]):
        """Sets the training_info of this TrainingInfo.


        :param training_info: The training_info of this TrainingInfo.
        :type training_info: List[TrainingInfoTrainingInfo]
        """

        self._training_info = training_info